Seperate subcommand structs and trait segregation to module
