+++ src/subcommand/mod.rs
@@ -1,0 +1,4 @@
+pub mod subcommand;
+pub mod this_subcommand;
+pub mod list_subcommand;
+pub mod generate_subcommand;
